+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - tt config is renamed to tt.yaml.
+- Do not use `make` command options for `tarantool` build if `MAKEFLAGS` environment variable
+is set.
 
 ### Fixed